refactor: move methods
