+++ src/schema.graphql
@@ -254,11 +254,11 @@ type Mutation {
   updateEventInfo(updateEventInfoInput: UpdateEventInfoInput!, queryId: ID!): EventQuery!
   addAttachments(attachments: [Upload!]!, queryId: ID!): EventQuery!
   deleteAttachments(attachmentIds: [ID!]!, queryId: ID!): EventQuery!
-  createQuery(createEventQueryInput: CreateEventQueryInput!): EventQuery!
-  expandQuery(expandEventQueryInput: ExpandEventQueryInput!, queryId: ID!): EventQuery!
   setQueryInviteToViewed(queryId: ID!): EventQuery!
   setQueryInviteToAccepted(queryId: ID!): EventQuery!
   setQueryInviteToDeclined(queryId: ID!): EventQuery!
+  createQuery(createEventQueryInput: CreateEventQueryInput!): EventQuery!
+  expandQuery(expandEventQueryInput: ExpandEventQueryInput!, queryId: ID!): EventQuery!
 }
 
 input UpdateBillingDetailsInput {
